pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["abini"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-cors-headers = "^4.3.1"
django-extensions = "^3.2.3"
django-rest-swagger = "^2.2.0"
django-sslserver = "^0.22"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"
mysqlclient = "^2.2.4"
reportlab = "^4.2.0"
yagmail = "^0.15.293"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
